CoffeeDelivery: Break Checkout page into smaller components
